Updates as discussed during review
* Have a dry_run mode instead of a list subcommand
* Validate that the resulting @SQ headers are as expected
* etc

+++ scripts/fix_sq_headers.py
@@ -1,12 +1,13 @@
 #!/usr/bin/env python3
 """
-Usage: fix_sq_headers list|update
+Usage: fix_sq_headers
 
-Uses the usual config entries (workflow.dataset/.access_level) and:
+Uses the usual config entries (workflow.dataset/.access_level/.dry_run) and:
 
   images.samtools      Path of an image containing samtools
   references.expected  Name (from KNOWN_REFS) of the (bad) refset expected
   references.new       Dict file containing the replacement @SQ headers
+  references.newset    Name (from KNOWN_REFS) of the replacement headers [optional]
 """
 
 import hashlib
@@ -38,6 +39,7 @@ query CramQuery($project: String!) {
 """)
 
 
+# Precomputed hashes (as per classify_SQ()) for common reference sets
 KNOWN_REFS = {
     '49f7a7583b832aed0104b1901f9cf5fd': 'Homo_sapiens_assembly38_masked',
     '760a471d32c9a356ffb37f632bebfea3': 'Homo_sapiens_assembly38[unmasked]',
@@ -50,12 +52,6 @@ def path_exists(path: Path):
     except Exception:
         # GSPath.exists() raises google.api_core.exceptions.NotFound if the bucket is missing
         return False
-
-
-def set_htslib_token():
-    gcloud_cmd = ['gcloud', 'auth', 'application-default', 'print-access-token']
-    token = sb.run(gcloud_cmd, check=True, capture_output=True, text=True).stdout
-    os.environ['GCS_OAUTH_TOKEN'] = token
 
 
 def parse_one_SQ(line: str):
@@ -84,7 +80,7 @@ def classify_SQ(path: str):
     return KNOWN_REFS.get(hashed, hashed)
 
 
-def do_reheader(refset, newref, incram, outcram, outcrai):
+def do_reheader(dry_run, refset, newref, newrefset, incram, outcram, outcrai):
     ref = classify_SQ(incram)
     if ref != refset:
         raise ValueError(f'{incram}: unexpected original reference headers: {ref}')
@@ -99,7 +95,7 @@ def do_reheader(refset, newref, incram, outcram, outcrai):
     head_cmd = ['samtools', 'head', incram]
     headers = sb.run(head_cmd, check=True, capture_output=True, text=True).stdout
 
-    newhdr_file = 'tmp.hdr'
+    newhdr_file = os.path.join(os.environ['BATCH_TMPDIR'], 'tmp.hdr')
     with open(newhdr_file, 'w') as newhdr:
         for line in headers.splitlines():
             if line.startswith('@SQ'):
@@ -112,9 +108,21 @@ def do_reheader(refset, newref, incram, outcram, outcrai):
             else:
                 print(line, file=newhdr)
 
+    if dry_run:
+        print(f'{incram} original headers: {ref}')
+        print(f'Would rewrite it to {outcram} with headers:')
+        with open(newhdr_file) as fp:
+            print(fp.read())
+        return
+
     sb.run(['samtools', 'reheader', '--no-PG', '--in-place', newhdr_file, incram], check=True)
     os.rename(incram, outcram)
     sb.run(['samtools', 'index', '-o', outcrai, outcram], check=True)
+
+    newref = classify_SQ(outcram)
+    print(f'{outcram}: reheadered: {newref}')
+    if newrefset is not None and newref != newrefset:
+        raise ValueError(f'{outcram}: unexpected reference headers written: {newref}')
 
 
 if __name__ == '__main__':
@@ -131,7 +139,7 @@ if __name__ == '__main__':
     for seqgroup in query(FIND_CRAMS, {'project': dataset})['project']['sequencingGroups']:
         for analysis in seqgroup['analyses']:
             fname = analysis['output']
-            if fname.endswith('.cram'):
+            if fname is not None and fname.endswith('.cram'):
                 cram_list.append(fname)
             else:
                 sgid = seqgroup['id']
@@ -140,56 +148,48 @@ if __name__ == '__main__':
 
     print(f'Total CRAM file database entries found: {len(cram_list)}')
 
-    command = sys.argv[1] if len(sys.argv) > 1 else 'help'
-    if command == 'list':
-        set_htslib_token()
-        for path in cram_list:
-            ref = classify_SQ(path) or '(file missing)'
-            print(f'{path}: {ref}')
+    expected_refset = config['references']['expected']
+    new_reference_path = config['references']['new']
+    new_refset = config['references'].get('newset', None)
+    dry_run = config['workflow'].get('dry_run', False)
 
-    elif command == 'update':
-        expected_refset = config['references']['expected']
-        new_reference_path = config['references']['new']
+    reheader_list = []
+    for fname in cram_list:
+        path = to_path(fname)
+        reheadered_path = path.parent / 'reheadered' / path.name
 
-        reheader_list = []
-        for fname in cram_list:
-            path = to_path(fname)
-            reheadered_path = path.parent / 'reheadered' / path.name
+        if not path_exists(path):
+            print(f'{fname}: file does not exist')
+        elif path_exists(reheadered_path):
+            print(f'{reheadered_path}: reheadered file already exists')
+        else:
+            reheader_list.append((path, reheadered_path, path.stat().st_size))
 
-            if not path_exists(path):
-                print(f'{fname}: file does not exist')
-            elif path_exists(reheadered_path):
-                print(f'{reheadered_path}: reheadered file already exists')
-            else:
-                reheader_list.append((path, reheadered_path, path.stat().st_size))
+    if len(reheader_list) == 0:
+        sys.exit('No CRAMs to be reheadered!')
 
-        if len(reheader_list) == 0:
-            sys.exit('No CRAMs to be reheadered!')
+    print(f'CRAM files to be reheadered: {len(reheader_list)}')
 
-        print(f'CRAM files to be reheadered: {len(reheader_list)}')
+    b = get_batch(f'Reheader {dataset} to {new_reference_path}')
+    new_reference = b.read_input(new_reference_path)
 
-        b = get_batch(f'Reheader {dataset} to {new_reference_path}')
-        new_reference = b.read_input(new_reference_path)
+    for path, newpath, filesize in reheader_list:
+        j = b.new_python_job(f'Reheader {path}')
+        j.image(image_path('samtools'))
+        j.storage(filesize * 1.2)  # Allow some extra space for index file, references, etc
 
-        for path, newpath, filesize in reheader_list:
-            j = b.new_python_job(f'Reheader {path}')
-            j.image(image_path('samtools'))
-            j.storage(filesize * 1.2)
+        j.call(do_reheader, dry_run, expected_refset, new_reference, new_refset,
+               b.read_input(str(path)), j.out_cram, j.out_crai)
 
-            j.call(do_reheader, expected_refset, new_reference,
-                   b.read_input(str(path)), j.out_cram, j.out_crai)
+        assert isinstance(j.out_cram, JobResourceFile)
+        assert isinstance(j.out_crai, JobResourceFile)
+        j.out_cram.add_extension('.cram')
+        j.out_crai.add_extension('.cram.crai')
 
-            assert isinstance(j.out_cram, JobResourceFile)
-            assert isinstance(j.out_crai, JobResourceFile)
-            j.out_cram.add_extension('.cram')
-            j.out_crai.add_extension('.cram.crai')
-
+        if not dry_run:
             b.write_output(j.out_cram, str(newpath))
             b.write_output(j.out_crai, str(newpath.with_suffix('.cram.crai')))
 
-            print(f'Added reheader job for {path}')
+        print(f'Added reheader job for {path}')
 
-        b.run(wait=False)
-
-    else:
-        sys.exit(__doc__)
+    b.run(wait=False)
